Use decimal in most places, add interest

+++ Grundkurs/Bankomat2/Transaction.cs
@@ -1,9 +1,9 @@
 namespace Bankomat2;
 
-class Transaction(bool isWithdraw, uint amount, DateTime timestamp)
+class Transaction(bool isWithdraw, decimal amount, DateTime timestamp)
 {
 
     public bool IsWithdraw { get; init; } = isWithdraw;
-    public uint Amount { get; init; } = amount;
+    public decimal Amount { get; init; } = amount;
     public DateTime Timestamp { get; init; } = timestamp;
 }
